Extract `FilterFunctionReturnTypeHelper`
